new prob seq before merging

+++ src/Bpp/Seq/Container/VectorPositionedContainer.h
@@ -241,7 +241,7 @@ namespace bpp
     void addObject_(std::shared_ptr<T> object, size_t objectIndex, bool checkPosition = false) const
     {
       if (objectIndex >= VectorPositionedContainer<T>::getSize())
-        throw IndexOutOfBoundsException("VectorPositionedContainer::addObject.", objectIndex, 0, VectorPositionedContainer<T>::getSize() - 1);
+        throw IndexOutOfBoundsException("VectorPositionedContainer::addObject.", objectIndex, 0, (int)VectorPositionedContainer<T>::getSize() - 1);
 
       if (checkPosition && positions_[objectIndex]!=nullptr)
         throw BadIntegerException("VectorPositionedContainer::setObject: object position already occupied in container ", (int)objectIndex);
